hw: Don't save branch result in RSS
We do not pass the branch result into the RS so keep the same RS implementation for
all FUs. For any write back we directly take the branch result from the FU. This is
possible because if we want to do a writeback the RS accepts the result only if the
writeback also accepts the writeback. Thus we can bypass the RS.

This should probably be solved in another way..