hw: Optimize timing of loop controller
The LEP iteration count was previously computed based on the FF update value and the again propagated to the RSS.
This value however depends on the actual execution.
The timing path was therefore from the loop controller, dispatcher, RSS, FU back via the RSS to the dispatcher & loop controller and again from there to the RSS.
We now compute it directly with dedicated hardware. This removes the last part from the loop controller to the RSS from the main path / it now can propagate in parallel.